requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic[email]
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
aiosqlite>=0.19.0
asyncpg>=0.28.0
passlib[bcrypt]
python-dotenv>=1.0.0
motor>=3.2.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0